コントローラーの整理
